:bug:Fix: STOMP 브로커 prefix 불일치 수정 및 WebSocket 설정 리팩토링
- /topic → /sub, /app → /pub prefix 통일
- /ws → /ws-stomp 엔드포인트 변경
- WebSocketConfig → WebSocketStompBrokerConfig 이름 변경 및 chat 도메인으로 이동
- ChatStompController 메시지 발송 경로 수정
